Update 19

See merge request ethereum-freeton-bridge/bridge-contracts!33

+++ ethereum/contracts/interfaces/IProxy.sol
@@ -10,7 +10,7 @@ interface IProxy {
         bytes eventData;
         int8 tonEventConfigurationWid;
         uint tonEventConfigurationAddress;
-        uint requiredConfirmations;
-        uint requiredRejects;
+        uint16 requiredConfirmations;
+        uint16 requiredRejects;
     }
 }
